worked on notifications 10:59
Added a map of customer2Notifications that will contain all the notifications for each customer.
In the forwarding yaz method, switch to add notification instead of simply moving the money between the accounts.
The notification is being added to the map.
Also I added an object named NotificationsDTO that will pass the notificationsd to the ui.

The ui can now call the method, pass as argument the customer name and receive all its notifications.

Need to add an update of the notifications in the initialize method and each time the tab 'payment' is being clicked.