Scramble
Started working on the featuer of valid form checking.
TODO:

create a boolean property that is binded to all the fields boolean properties.
Check that when all the fields are valid the form validation boolean property is set to true, otherwise it is set to false.

implement the methods that actually checks the validity of each field (regex).

final check that when one of the fields isn't valid the validiation of the form isn't valid.

Then, bind the validition of the form to the scramble button.

For later - add the css classes for the invalid text, so they will be colored in red or something like that..

Then - implement the DTO creation and sending to model. Check that the loan placment works fine.

Then - watch the final example and create a task of a new thread in the model, also create the loading panel.